+++ openfisca_survey_manager/config.py
@@ -9,7 +9,7 @@ class Config(configparser.ConfigParser):
         configparser.ConfigParser.__init__(self)
         if config_files_directory is not None:
             config_ini = os.path.join(config_files_directory, 'config.ini')
-            assert os.path.exists(config_ini), "{} is not a valid path".format(config_ini)
+            assert os.path.exists(config_ini), f"{config_ini} is not a valid path"
             self.config_ini = config_ini
             self.read([config_ini])